Moved metadata methods to image services.
